backend/requirements.txt
fastapi
uvicorn
supabase
httpx
python-dotenv
joblib
pydantic
pydantic-settings
python-multipart
websockets
